Moves the constant to a variable in config folder
